src/requirements.txt
argparse-addons==0.12.0
bitstruct==8.20.0
cantools==40.2.1
crccheck==1.3.0
diskcache==5.6.3
msgpack==1.1.0
packaging==24.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-can==4.5.0
textparser==0.24.0
typing_extensions==4.13.2
wrapt==1.17.2
